[asset-mapping] add asset-mapping project
