+++ asset-mapping/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "starcoin-asset-mapping"
+version = "2.0.1"
+authors = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+
+[dependencies]
+anyhow = { workspace = true }
+bcs-ext = { workspace = true }
+test-helper = { workspace = true }
+starcoin-crypto = { workspace = true }
+starcoin-types = { workspace = true }
+starcoin-state-api = { workspace = true }
+starcoin-vm-types = { workspace = true }
+starcoin-config = { workspace = true }
+forkable-jellyfish-merkle = { workspace = true }
